+++ backend/gce.go
@@ -55,6 +55,8 @@ const (
 	defaultGCEUploadRetrySleep   = 1 * time.Second
 	defaultGCEImageSelectorType  = "env"
 	defaultGCEImage              = "travis-ci.+"
+	defaultGCEGpuCount           = int64(0)
+	defaultGCEGpuType            = "nvidia-tesla-p100"
 	defaultGCERateLimitMaxCalls  = uint64(10)
 	defaultGCERateLimitDuration  = time.Second
 	defaultGCESSHDialTimeout     = 5 * time.Second
@@ -69,6 +71,8 @@ var (
 		"DEFAULT_LANGUAGE":       fmt.Sprintf("default language to use when looking up image (default %q)", defaultGCELanguage),
 		"DETERMINISTIC_HOSTNAME": "assign deterministic hostname based on repo slug and job id (default false)",
 		"DISK_SIZE":              fmt.Sprintf("disk size in GB (default %v)", defaultGCEDiskSize),
+		"GPU_COUNT":              fmt.Sprintf("number of GPUs to use (default %v)", defaultGCEGpuCount),
+		"GPU_TYPE":               fmt.Sprintf("type of GPU to use (default %q)", defaultGCEGpuType),
 		"IMAGE_ALIASES":          "comma-delimited strings used as stable names for images, used only when image selector type is \"env\"",
 		"IMAGE_DEFAULT":          fmt.Sprintf("default image name to use when none found (default %q)", defaultGCEImage),
 		"IMAGE_SELECTOR_TYPE":    fmt.Sprintf("image selector type (\"env\" or \"api\", default %q)", defaultGCEImageSelectorType),
@@ -156,6 +160,8 @@ type gceProvider struct {
 	bootPrePollSleep      time.Duration
 	defaultLanguage       string
 	defaultImage          string
+	defaultGpuCount       int64
+	defaultGpuType        string
 	uploadRetries         uint64
 	uploadRetrySleep      time.Duration
 	sshDialer             ssh.Dialer
@@ -173,6 +179,7 @@ type gceInstanceConfig struct {
 	Zone               *compute.Zone
 	Network            *compute.Network
 	Subnetwork         *compute.Subnetwork
+	AcceleratorConfig  *compute.AcceleratorConfig
 	DiskType           string
 	DiskSize           int64
 	SSHPubKey          string
@@ -199,22 +206,27 @@ func (gsmw *gceStartMultistepWrapper) Run(multistep.StateBag) multistep.StepActi
 func (gsmw *gceStartMultistepWrapper) Cleanup(multistep.StateBag) { return }
 
 type gceStartContext struct {
-	startAttributes  *StartAttributes
-	ctx              gocontext.Context
-	instChan         chan Instance
-	errChan          chan error
-	image            *compute.Image
-	script           string
-	bootStart        time.Time
-	instance         *compute.Instance
-	instanceInsertOp *compute.Operation
+	startAttributes      *StartAttributes
+	progresser           Progresser
+	ctx                  gocontext.Context
+	instChan             chan Instance
+	errChan              chan error
+	image                *compute.Image
+	script               string
+	bootStart            time.Time
+	instance             *compute.Instance
+	instanceInsertOpName string
+	zoneName             string
 }
 
 type gceInstance struct {
+	zoneName string
 	client   *compute.Service
 	provider *gceProvider
 	instance *compute.Instance
 	ic       *gceInstanceConfig
+
+	progresser Progresser
 
 	authUser     string
 	cachedIPAddr string
@@ -377,6 +389,22 @@ func newGCEProvider(cfg *config.ProviderConfig) (Provider, error) {
 		defaultImage = cfg.Get("IMAGE_DEFAULT")
 	}
 
+	defaultAcceleratorConfig := &compute.AcceleratorConfig{}
+
+	defaultAcceleratorConfig.AcceleratorType = defaultGCEGpuType
+	if cfg.IsSet("GPU_TYPE") {
+		defaultAcceleratorConfig.AcceleratorType = cfg.Get("GPU_TYPE")
+	}
+
+	defaultAcceleratorConfig.AcceleratorCount = defaultGCEGpuCount
+	if cfg.IsSet("GPU_COUNT") {
+		dgc, err := strconv.ParseInt(cfg.Get("GPU_COUNT"), 0, 64)
+		if err != nil {
+			return nil, err
+		}
+		defaultAcceleratorConfig.AcceleratorCount = dgc
+	}
+
 	autoImplode := true
 	if cfg.IsSet("AUTO_IMPLODE") {
 		ai, err := strconv.ParseBool(cfg.Get("AUTO_IMPLODE"))
@@ -477,16 +505,17 @@ func newGCEProvider(cfg *config.ProviderConfig) (Provider, error) {
 		sshDialTimeout: sshDialTimeout,
 
 		ic: &gceInstanceConfig{
-			Preemptible:      preemptible,
-			PublicIP:         publicIP,
-			PublicIPConnect:  publicIPConnect,
-			DiskSize:         diskSize,
-			SSHPubKey:        string(pubKey),
-			AutoImplode:      autoImplode,
-			StopPollSleep:    stopPollSleep,
-			StopPrePollSleep: stopPrePollSleep,
-			SkipStopPoll:     skipStopPoll,
-			Site:             site,
+			Preemptible:       preemptible,
+			PublicIP:          publicIP,
+			PublicIPConnect:   publicIPConnect,
+			DiskSize:          diskSize,
+			SSHPubKey:         string(pubKey),
+			AutoImplode:       autoImplode,
+			StopPollSleep:     stopPollSleep,
+			StopPrePollSleep:  stopPrePollSleep,
+			SkipStopPoll:      skipStopPoll,
+			Site:              site,
+			AcceleratorConfig: defaultAcceleratorConfig,
 		},
 
 		deterministicHostname: deterministicHostname,
@@ -633,13 +662,33 @@ func loadGoogleAccountJSON(filenameOrJSON string) (*gceAccountJSON, error) {
 	return a, err
 }
 
-func (p *gceProvider) Start(ctx gocontext.Context, startAttributes *StartAttributes) (Instance, error) {
+func (p *gceProvider) SupportsProgress() bool {
+	return true
+}
+
+func (p *gceProvider) StartWithProgress(ctx gocontext.Context, startAttributes *StartAttributes, progresser Progresser) (Instance, error) {
 	logger := context.LoggerFromContext(ctx).WithField("self", "backend/gce_provider")
+
+	var (
+		zone *compute.Zone
+		err  error
+	)
+	zone = p.ic.Zone
+	if startAttributes.VMConfig.Zone != "" {
+		logger.WithField("zone", startAttributes.VMConfig.Zone).Debug("setting zone from vm config")
+
+		zone, err = p.client.Zones.Get(p.projectID, startAttributes.VMConfig.Zone).Context(ctx).Do()
+		if err != nil {
+			return nil, err
+		}
+	}
 
 	state := &multistep.BasicStateBag{}
 
 	c := &gceStartContext{
 		startAttributes: startAttributes,
+		zoneName:        zone.Name,
+		progresser:      progresser,
 		ctx:             ctx,
 		instChan:        make(chan Instance),
 		errChan:         make(chan error),
@@ -659,7 +708,7 @@ func (p *gceProvider) Start(ctx gocontext.Context, startAttributes *StartAttribu
 	defer func(c *gceStartContext) {
 		if c.instance != nil && abandonedStart {
 			p.apiRateLimit(c.ctx)
-			_, _ = p.client.Instances.Delete(p.projectID, p.ic.Zone.Name, c.instance.Name).Do()
+			_, _ = p.client.Instances.Delete(p.projectID, c.zoneName, c.instance.Name).Do()
 		}
 	}(c)
 
@@ -676,19 +725,36 @@ func (p *gceProvider) Start(ctx gocontext.Context, startAttributes *StartAttribu
 		if ctx.Err() == gocontext.DeadlineExceeded {
 			metrics.Mark("worker.vm.provider.gce.boot.timeout")
 		}
+		c.progresser.Progress(&ProgressEntry{
+			Message:    "timeout waiting for instance to be ready",
+			State:      ProgressFailure,
+			Interrupts: true,
+		})
 		abandonedStart = true
 		return nil, ctx.Err()
 	}
 }
 
+func (p *gceProvider) Start(ctx gocontext.Context, startAttributes *StartAttributes) (Instance, error) {
+	return p.StartWithProgress(ctx, startAttributes, NewTextProgresser(ioutil.Discard))
+}
+
 func (p *gceProvider) stepGetImage(c *gceStartContext) multistep.StepAction {
 	image, err := p.imageSelect(c.ctx, c.startAttributes)
 	if err != nil {
+		c.progresser.Progress(&ProgressEntry{
+			Message: "could not select image",
+			State:   ProgressFailure,
+		})
 		c.errChan <- err
 		return multistep.ActionHalt
 	}
 
 	c.image = image
+	c.progresser.Progress(&ProgressEntry{
+		Message: fmt.Sprintf("selected image %q", image.Name),
+		State:   ProgressSuccess,
+	})
 	return multistep.ActionContinue
 }
 
@@ -701,11 +767,19 @@ func (p *gceProvider) stepRenderScript(c *gceStartContext) multistep.StepAction 
 	}
 	err := gceStartupScript.Execute(&scriptBuf, scriptData)
 	if err != nil {
+		c.progresser.Progress(&ProgressEntry{
+			Message: "could not render startup script",
+			State:   ProgressFailure,
+		})
 		c.errChan <- err
 		return multistep.ActionHalt
 	}
 
 	c.script = scriptBuf.String()
+	c.progresser.Progress(&ProgressEntry{
+		Message: "rendered startup script",
+		State:   ProgressSuccess,
+	})
 	return multistep.ActionContinue
 }
 
@@ -720,14 +794,22 @@ func (p *gceProvider) stepInsertInstance(c *gceStartContext) multistep.StepActio
 	c.bootStart = time.Now().UTC()
 
 	p.apiRateLimit(c.ctx)
-	op, err := p.client.Instances.Insert(p.projectID, p.ic.Zone.Name, inst).Context(c.ctx).Do()
+	op, err := p.client.Instances.Insert(p.projectID, c.zoneName, inst).Context(c.ctx).Do()
 	if err != nil {
+		c.progresser.Progress(&ProgressEntry{
+			Message: "could not insert instance",
+			State:   ProgressFailure,
+		})
 		c.errChan <- err
 		return multistep.ActionHalt
 	}
 
 	c.instance = inst
-	c.instanceInsertOp = op
+	c.instanceInsertOpName = op.Name
+	c.progresser.Progress(&ProgressEntry{
+		Message: "inserted instance",
+		State:   ProgressSuccess,
+	})
 	return multistep.ActionContinue
 }
 
@@ -735,44 +817,71 @@ func (p *gceProvider) stepWaitForInstanceIP(c *gceStartContext) multistep.StepAc
 	logger := context.LoggerFromContext(c.ctx).WithField("self", "backend/gce_provider")
 
 	logger.WithField("duration", p.bootPrePollSleep).Debug("sleeping before first checking instance insert operation")
+	c.progresser.Progress(&ProgressEntry{
+		Message: fmt.Sprintf("sleeping %s before checking instance insert", p.bootPrePollSleep),
+		State:   ProgressNeutral,
+	})
 
 	time.Sleep(p.bootPrePollSleep)
 
-	zoneOpCall := p.client.ZoneOperations.Get(p.projectID, p.ic.Zone.Name, c.instanceInsertOp.Name).Context(c.ctx)
+	zoneOpCall := p.client.ZoneOperations.Get(p.projectID, c.zoneName, c.instanceInsertOpName).Context(c.ctx)
 
+	c.progresser.Progress(&ProgressEntry{
+		Message:   "polling for instance insert completion...",
+		State:     ProgressNeutral,
+		Continues: true,
+	})
 	for {
 		metrics.Mark("worker.vm.provider.gce.boot.poll")
 
 		p.apiRateLimit(c.ctx)
 		newOp, err := zoneOpCall.Do()
 		if err != nil {
+			c.progresser.Progress(&ProgressEntry{
+				Message:    "could not check for instance insert",
+				State:      ProgressFailure,
+				Interrupts: true,
+			})
 			c.errChan <- err
 			return multistep.ActionHalt
 		}
 
 		if newOp.Status == "RUNNING" || newOp.Status == "DONE" {
 			if newOp.Error != nil {
+				c.progresser.Progress(&ProgressEntry{
+					Message:    "instance could not be inserted",
+					State:      ProgressFailure,
+					Interrupts: true,
+				})
 				c.errChan <- &gceOpError{Err: newOp.Error}
 				return multistep.ActionHalt
 			}
 
 			logger.WithFields(logrus.Fields{
 				"status": newOp.Status,
-				"name":   c.instanceInsertOp.Name,
+				"name":   c.instanceInsertOpName,
 			}).Debug("instance is ready")
 
+			startupDuration := time.Now().UTC().Sub(c.bootStart)
+			c.progresser.Progress(&ProgressEntry{
+				Message:    fmt.Sprintf("instance is ready (%s)", startupDuration.Truncate(time.Millisecond)),
+				State:      ProgressSuccess,
+				Interrupts: true,
+			})
 			c.instChan <- &gceInstance{
-				client:   p.client,
-				provider: p,
-				instance: c.instance,
-				ic:       p.ic,
+				zoneName:   c.zoneName,
+				client:     p.client,
+				provider:   p,
+				instance:   c.instance,
+				ic:         p.ic,
+				progresser: c.progresser,
 
 				authUser: "travis",
 
 				projectID: p.projectID,
 				imageName: c.image.Name,
 
-				startupDuration: time.Now().UTC().Sub(c.bootStart),
+				startupDuration: startupDuration,
 			}
 			return multistep.ActionContinue
 		}
@@ -780,19 +889,25 @@ func (p *gceProvider) stepWaitForInstanceIP(c *gceStartContext) multistep.StepAc
 		if newOp.Error != nil {
 			logger.WithFields(logrus.Fields{
 				"err":  newOp.Error,
-				"name": c.instanceInsertOp.Name,
+				"name": c.instanceInsertOpName,
 			}).Error("encountered an error while waiting for instance insert operation")
 
+			c.progresser.Progress(&ProgressEntry{
+				Message:    "error while waiting for instance insert",
+				State:      ProgressFailure,
+				Interrupts: true,
+			})
 			c.errChan <- &gceOpError{Err: newOp.Error}
 			return multistep.ActionHalt
 		}
 
 		logger.WithFields(logrus.Fields{
 			"status":   newOp.Status,
-			"name":     c.instanceInsertOp.Name,
+			"name":     c.instanceInsertOpName,
 			"duration": p.bootPollSleep,
 		}).Debug("sleeping before checking instance insert operation")
 
+		c.progresser.Progress(&ProgressEntry{Message: ".", Raw: true})
 		time.Sleep(p.bootPollSleep)
 	}
 }
@@ -872,12 +987,50 @@ func buildGCEImageSelector(selectorType string, cfg *config.ProviderConfig) (ima
 }
 
 func (p *gceProvider) buildInstance(ctx gocontext.Context, startAttributes *StartAttributes, imageLink, startupScript string) *compute.Instance {
+	logger := context.LoggerFromContext(ctx).WithField("self", "backend/gce_instance")
+
+	var err error
+
+	zone := p.ic.Zone
+	diskType := p.ic.DiskType
+
+	if startAttributes.VMConfig.Zone != "" {
+		zone, err = p.client.Zones.Get(p.projectID, startAttributes.VMConfig.Zone).Context(ctx).Do()
+		if err != nil {
+			logger.Warn(err)
+		}
+		logger.WithFields(logrus.Fields{
+			"disk_type": fmt.Sprintf("zones/%s/diskTypes/pd-ssd", zone.Name),
+		}).Info("setting disk type based on zone in vm config")
+		diskType = fmt.Sprintf("zones/%s/diskTypes/pd-ssd", zone.Name)
+
+	}
+
+	/* TODO: We should do this only if we're using a custom zone to avoid unnecessary lookups. */
 	var machineType *compute.MachineType
-	switch startAttributes.VMType {
-	case "premium":
-		machineType = p.ic.PremiumMachineType
-	default:
-		machineType = p.ic.MachineType
+	if startAttributes.VMType == "premium" {
+		pic, err := p.client.MachineTypes.Get(p.projectID, zone.Name, p.cfg.Get("PREMIUM_MACHINE_TYPE")).Context(ctx).Do()
+		if err != nil {
+			logger.WithField("err", err).Warn("failed to look up premium machine type")
+		}
+		machineType = pic
+	} else {
+		p.apiRateLimit(ctx)
+		pic, err := p.client.MachineTypes.Get(p.projectID, zone.Name, p.cfg.Get("MACHINE_TYPE")).Context(ctx).Do()
+		if err != nil {
+			logger.WithField("err", err).Warn("failed to look up machine type")
+		}
+		machineType = pic
+	}
+
+	// Set accelerator config based on number and type of requested GPUs (empty if none)
+	acceleratorConfig := &compute.AcceleratorConfig{}
+	if startAttributes.VMConfig.GpuCount > 0 {
+		acceleratorConfig.AcceleratorCount = startAttributes.VMConfig.GpuCount
+		acceleratorConfig.AcceleratorType = fmt.Sprintf("https://www.googleapis.com/compute/v1/projects/%s/zones/%s/acceleratorTypes/%s",
+			p.projectID,
+			startAttributes.VMConfig.Zone,
+			startAttributes.VMConfig.GpuType)
 	}
 
 	var subnetwork string
@@ -917,6 +1070,12 @@ func (p *gceProvider) buildInstance(ctx gocontext.Context, startAttributes *Star
 		hostname = fmt.Sprintf("travis-job-%s", uuid.NewRandom())
 	}
 
+	acceleratorConfigs := []*compute.AcceleratorConfig{}
+	if acceleratorConfig.AcceleratorCount > 0 {
+		logger.Debug("GPU requested, setting acceleratorConfig")
+		acceleratorConfigs = append(acceleratorConfigs, acceleratorConfig)
+	}
+
 	return &compute.Instance{
 		Description: fmt.Sprintf("Travis CI %s test VM", startAttributes.Language),
 		Disks: []*compute.AttachedDisk{
@@ -927,14 +1086,16 @@ func (p *gceProvider) buildInstance(ctx gocontext.Context, startAttributes *Star
 				AutoDelete: true,
 				InitializeParams: &compute.AttachedDiskInitializeParams{
 					SourceImage: imageLink,
-					DiskType:    p.ic.DiskType,
+					DiskType:    diskType,
 					DiskSizeGb:  p.ic.DiskSize,
 				},
 			},
 		},
+		GuestAccelerators: acceleratorConfigs,
 		Scheduling: &compute.Scheduling{
-			Preemptible:      p.ic.Preemptible,
-			AutomaticRestart: googleapi.Bool(false),
+			Preemptible:       p.ic.Preemptible,
+			AutomaticRestart:  googleapi.Bool(false),
+			OnHostMaintenance: "TERMINATE",
 		},
 		MachineType: machineType.SelfLink,
 		Name:        hostname,
@@ -999,7 +1160,7 @@ func (i *gceInstance) getIP() string {
 
 func (i *gceInstance) refreshInstance(ctx gocontext.Context) error {
 	i.provider.apiRateLimit(ctx)
-	inst, err := i.client.Instances.Get(i.projectID, i.ic.Zone.Name, i.instance.Name).Context(ctx).Do()
+	inst, err := i.client.Instances.Get(i.projectID, i.zoneName, i.instance.Name).Context(ctx).Do()
 	if err != nil {
 		return err
 	}
@@ -1008,9 +1169,20 @@ func (i *gceInstance) refreshInstance(ctx gocontext.Context) error {
 	return nil
 }
 
+func (i *gceInstance) SupportsProgress() bool {
+	return true
+}
+
 func (i *gceInstance) UploadScript(ctx gocontext.Context, script []byte) error {
 	uploadedChan := make(chan error)
 	var lastErr error
+
+	waitStart := time.Now().UTC()
+	i.progresser.Progress(&ProgressEntry{
+		Message:   "waiting for ssh connectivity...",
+		State:     ProgressNeutral,
+		Continues: true,
+	})
 
 	go func() {
 		var errCount uint64
@@ -1021,6 +1193,16 @@ func (i *gceInstance) UploadScript(ctx gocontext.Context, script []byte) error {
 
 			err := i.uploadScriptAttempt(ctx, script)
 			if err == nil {
+				timeToSsh := time.Now().UTC().Sub(waitStart).Truncate(time.Millisecond)
+				i.progresser.Progress(&ProgressEntry{
+					Message:    fmt.Sprintf("ssh connectivity established (%s)", timeToSsh),
+					State:      ProgressSuccess,
+					Interrupts: true,
+				})
+				i.progresser.Progress(&ProgressEntry{
+					Message: "uploaded script",
+					State:   ProgressSuccess,
+				})
 				uploadedChan <- nil
 				return
 			}
@@ -1033,6 +1215,7 @@ func (i *gceInstance) UploadScript(ctx gocontext.Context, script []byte) error {
 				return
 			}
 
+			i.progresser.Progress(&ProgressEntry{Message: ".", Raw: true})
 			time.Sleep(i.provider.uploadRetrySleep)
 		}
 	}()
@@ -1058,6 +1241,11 @@ func (i *gceInstance) uploadScriptAttempt(ctx gocontext.Context, script []byte) 
 
 	existed, err := conn.UploadFile("build.sh", script)
 	if existed {
+		i.progresser.Progress(&ProgressEntry{
+			Message:    "existing script detected",
+			State:      ProgressFailure,
+			Interrupts: true,
+		})
 		return ErrStaleVM
 	}
 	if err != nil {
@@ -1129,6 +1317,44 @@ func (i *gceInstance) RunScript(ctx gocontext.Context, output io.Writer) (*RunRe
 	return &RunResult{Completed: err != nil, ExitCode: exitStatus}, errors.Wrap(err, "error running script")
 }
 
+func (i *gceInstance) DownloadTrace(ctx gocontext.Context) ([]byte, error) {
+	var buf []byte
+	var err error
+
+	downloadedChan := make(chan error)
+
+	go func() {
+		buf, err = i.downloadTrace(ctx)
+		downloadedChan <- err
+	}()
+
+	select {
+	case err := <-downloadedChan:
+		return buf, err
+	case <-ctx.Done():
+		context.LoggerFromContext(ctx).WithFields(logrus.Fields{
+			"err":  ctx.Err(),
+			"self": "backend/gce_instance",
+		}).Info("stopping upload retries, error from last attempt")
+		return nil, ctx.Err()
+	}
+}
+
+func (i *gceInstance) downloadTrace(ctx gocontext.Context) ([]byte, error) {
+	conn, err := i.sshConnection(ctx)
+	if err != nil {
+		return nil, errors.Wrap(err, "couldn't connect to SSH server")
+	}
+	defer conn.Close()
+
+	buf, err := conn.DownloadFile("/tmp/build.trace")
+	if err != nil {
+		return nil, errors.Wrap(err, "couldn't download trace")
+	}
+
+	return buf, nil
+}
+
 func (i *gceInstance) Stop(ctx gocontext.Context) error {
 	logger := context.LoggerFromContext(ctx).WithField("self", "backend/gce_instance")
 	state := &multistep.BasicStateBag{}
@@ -1161,7 +1387,7 @@ func (i *gceInstance) Stop(ctx gocontext.Context) error {
 }
 
 func (i *gceInstance) stepDeleteInstance(c *gceInstanceStopContext) multistep.StepAction {
-	op, err := i.client.Instances.Delete(i.projectID, i.ic.Zone.Name, i.instance.Name).Do()
+	op, err := i.client.Instances.Delete(i.projectID, i.zoneName, i.instance.Name).Do()
 	if err != nil {
 		c.errChan <- err
 		return multistep.ActionHalt
@@ -1187,7 +1413,7 @@ func (i *gceInstance) stepWaitForInstanceDeleted(c *gceInstanceStopContext) mult
 	time.Sleep(i.ic.StopPrePollSleep)
 
 	zoneOpCall := i.client.ZoneOperations.Get(i.projectID,
-		i.ic.Zone.Name, c.instanceDeleteOp.Name)
+		i.zoneName, c.instanceDeleteOp.Name)
 
 	b := backoff.NewExponentialBackOff()
 	b.InitialInterval = i.ic.StopPollSleep