+++ backend/package.go
@@ -69,6 +69,13 @@ type Provider interface {
 	// ready to call UploadScript on (this may, for example, mean that it
 	// waits for SSH connections to be possible).
 	Start(gocontext.Context, *StartAttributes) (Instance, error)
+
+	// StartWithProgress starts an instance as with Start and also reports
+	// progress via an io.Writer.
+	StartWithProgress(gocontext.Context, *StartAttributes, Progresser) (Instance, error)
+
+	// SupportsProgress allows for querying of progress support, yeah!
+	SupportsProgress() bool
 }
 
 // An Instance is something that can run a build script.
@@ -81,6 +88,11 @@ type Instance interface {
 	// RunScript runs the build script that was uploaded with the
 	// UploadScript method.
 	RunScript(gocontext.Context, io.Writer) (*RunResult, error)
+
+	// DownloadTrace attempts to download a job trace from the instance
+	DownloadTrace(gocontext.Context) ([]byte, error)
+
+	// Stop stops (and deletes) the instance
 	Stop(gocontext.Context) error
 
 	// ID is used when identifying the instance in logs and such
@@ -91,6 +103,9 @@ type Instance interface {
 
 	// StartupDuration is the duration between "created" and "ready"
 	StartupDuration() time.Duration
+
+	// SupportsProgress allows for querying of progress support, yeah!
+	SupportsProgress() bool
 }
 
 // RunResult represents the result of running a script with Instance.RunScript.