+++ backend/start_attributes_test.go
@@ -20,6 +20,7 @@ var (
 				{Language: "python", Dist: "frob", Group: "edge", OS: "flob"},
 				{Language: "python", Dist: "frob", OsxImage: "", Group: "edge", OS: "flob"},
 				{Language: "python", Dist: "frob", OsxImage: "", Group: "edge", OS: "flob", VMType: "premium"},
+				{Language: "python", Dist: "frob", OsxImage: "", Group: "edge", OS: "flob", VMType: "premium", VMConfig: VmConfig{GpuCount: 0, GpuType: "", Zone: ""}},
 			},
 			O: []*StartAttributes{
 				{Language: "default", Dist: "precise", Group: "stable", OS: "linux", VMType: "default"},
@@ -29,6 +30,7 @@ var (
 				{Language: "python", Dist: "frob", Group: "edge", OS: "flob", VMType: "default"},
 				{Language: "python", Dist: "frob", OsxImage: "", Group: "edge", OS: "flob", VMType: "default"},
 				{Language: "python", Dist: "frob", OsxImage: "", Group: "edge", OS: "flob", VMType: "premium"},
+				{Language: "python", Dist: "frob", OsxImage: "", Group: "edge", OS: "flob", VMType: "premium", VMConfig: VmConfig{}},
 			},
 		},
 	}
@@ -42,13 +44,14 @@ func TestStartAttributes(t *testing.T) {
 	assert.Equal(t, "", sa.OS)
 	assert.Equal(t, "", sa.OsxImage)
 	assert.Equal(t, "", sa.VMType)
+	assert.Equal(t, VmConfig{GpuCount: 0, GpuType: "", Zone: ""}, sa.VMConfig)
 }
 
 func TestStartAttributes_SetDefaults(t *testing.T) {
 	for _, tc := range testStartAttributesTestCases {
 		for i, sa := range tc.A {
 			expected := tc.O[i]
-			sa.SetDefaults("default", "precise", "stable", "linux", "default")
+			sa.SetDefaults("default", "precise", "stable", "linux", "default", sa.VMConfig)
 			assert.Equal(t, expected, sa)
 		}
 	}