+++ cli.go
@@ -52,11 +52,12 @@ type CLI struct {
 
 	Config                  *config.Config
 	BuildScriptGenerator    BuildScriptGenerator
+	BuildTracePersister     BuildTracePersister
 	BackendProvider         backend.Provider
 	ProcessorPool           *ProcessorPool
 	CancellationBroadcaster *CancellationBroadcaster
 	JobQueue                JobQueue
-	LogsQueue               LogsQueue
+	LogWriterFactory        LogWriterFactory
 
 	heartbeatErrSleep time.Duration
 	heartbeatSleep    time.Duration
@@ -138,6 +139,11 @@ func (i *CLI) Setup() (bool, error) {
 
 	i.BuildScriptGenerator = generator
 
+	persister := NewBuildTracePersister(i.Config)
+	logger.WithField("build_trace_persister", fmt.Sprintf("%#v", persister)).Debug("built")
+
+	i.BuildTracePersister = persister
+
 	if i.Config.TravisSite != "" {
 		i.Config.ProviderConfig.Set("TRAVIS_SITE", i.Config.TravisSite)
 	}
@@ -171,7 +177,7 @@ func (i *CLI) Setup() (bool, error) {
 		PayloadFilterExecutable: i.Config.PayloadFilterExecutable,
 	}
 
-	pool := NewProcessorPool(ppc, i.BackendProvider, i.BuildScriptGenerator, i.CancellationBroadcaster)
+	pool := NewProcessorPool(ppc, i.BackendProvider, i.BuildScriptGenerator, i.BuildTracePersister, i.CancellationBroadcaster)
 
 	pool.SkipShutdownOnLogTimeout = i.Config.SkipShutdownOnLogTimeout
 	logger.WithField("pool", pool).Debug("built")
@@ -184,7 +190,7 @@ func (i *CLI) Setup() (bool, error) {
 		return false, err
 	}
 
-	err = i.setupLogsQueue()
+	err = i.setupLogWriterFactory()
 	if err != nil {
 		logger.WithField("err", err).Error("couldn't create logs queue")
 		return false, err
@@ -211,20 +217,20 @@ func (i *CLI) Run() {
 	go i.signalHandler()
 
 	i.logger.WithFields(logrus.Fields{
-		"pool_size":  i.Config.PoolSize,
-		"queue":      i.JobQueue,
-		"logs_queue": i.LogsQueue,
+		"pool_size":         i.Config.PoolSize,
+		"queue":             i.JobQueue,
+		"logwriter_factory": i.LogWriterFactory,
 	}).Debug("running pool")
 
-	i.ProcessorPool.Run(i.Config.PoolSize, i.JobQueue)
+	i.ProcessorPool.Run(i.Config.PoolSize, i.JobQueue, i.LogWriterFactory)
 
 	err := i.JobQueue.Cleanup()
 	if err != nil {
 		i.logger.WithField("err", err).Error("couldn't clean up job queue")
 	}
 
-	if i.LogsQueue != nil {
-		err := i.LogsQueue.Cleanup()
+	if i.LogWriterFactory != nil {
+		err := i.LogWriterFactory.Cleanup()
 		if err != nil {
 			i.logger.WithField("err", err).Error("couldn't clean up logs queue")
 		}
@@ -620,14 +626,26 @@ func (i *CLI) buildAMQPJobQueueAndCanceller() (*AMQPJobQueue, *AMQPCanceller, er
 			}
 			cfg.RootCAs.AppendCertsFromPEM(cert)
 		}
-		amqpConn, err = amqp.DialTLS(i.Config.AmqpURI, cfg)
+		amqpConn, err = amqp.DialConfig(i.Config.AmqpURI,
+			amqp.Config{
+				Heartbeat:       i.Config.AmqpHeartbeat,
+				Locale:          "en_US",
+				TLSClientConfig: cfg,
+			})
 	} else if i.Config.AmqpInsecure {
-		amqpConn, err = amqp.DialTLS(
+		amqpConn, err = amqp.DialConfig(
 			i.Config.AmqpURI,
-			&tls.Config{InsecureSkipVerify: true},
-		)
+			amqp.Config{
+				Heartbeat:       i.Config.AmqpHeartbeat,
+				Locale:          "en_US",
+				TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+			})
 	} else {
-		amqpConn, err = amqp.Dial(i.Config.AmqpURI)
+		amqpConn, err = amqp.DialConfig(i.Config.AmqpURI,
+			amqp.Config{
+				Heartbeat: i.Config.AmqpHeartbeat,
+				Locale:    "en_US",
+			})
 	}
 	if err != nil {
 		i.logger.WithField("err", err).Error("couldn't connect to AMQP")
@@ -641,7 +659,12 @@ func (i *CLI) buildAMQPJobQueueAndCanceller() (*AMQPJobQueue, *AMQPCanceller, er
 	canceller := NewAMQPCanceller(i.ctx, amqpConn, i.CancellationBroadcaster)
 	i.logger.WithField("canceller", fmt.Sprintf("%#v", canceller)).Debug("built")
 
-	jobQueue, err := NewAMQPJobQueue(amqpConn, i.Config.QueueName, i.Config.StateUpdatePoolSize)
+	jobQueue, err := NewAMQPJobQueue(amqpConn, i.Config.QueueName, i.Config.StateUpdatePoolSize, i.Config.RabbitMQSharding)
+
+	// Set the consumer priority directly instead of altering the signature of
+	// NewAMQPJobQueue :sigh_cat:
+	jobQueue.priority = i.Config.AmqpConsumerPriority
+
 	if err != nil {
 		return nil, nil, err
 	}
@@ -692,16 +715,20 @@ func (i *CLI) buildFileJobQueue() (*FileJobQueue, error) {
 	return jobQueue, nil
 }
 
-func (i *CLI) setupLogsQueue() error {
+func (i *CLI) setupLogWriterFactory() error {
 	if i.Config.LogsAmqpURI == "" {
 		// If no separate URI is set for LogsAMQP, use the JobsQueue to send log parts
 		return nil
 	}
-	err := i.buildAMQPLogsQueue()
-	return err
+	logWriterFactory, err := i.buildAMQPLogWriterFactory()
+	if err != nil {
+		return err
+	}
+	i.LogWriterFactory = logWriterFactory
+	return nil
 }
 
-func (i *CLI) buildAMQPLogsQueue() error {
+func (i *CLI) buildAMQPLogWriterFactory() (*AMQPLogWriterFactory, error) {
 	var amqpConn *amqp.Connection
 	var err error
 
@@ -714,27 +741,44 @@ func (i *CLI) buildAMQPLogsQueue() error {
 		if i.Config.LogsAmqpTlsCertPath != "" {
 			cert, err := ioutil.ReadFile(i.Config.LogsAmqpTlsCertPath)
 			if err != nil {
-				return err
+				return nil, err
 			}
 			cfg.RootCAs.AppendCertsFromPEM(cert)
 		}
-		amqpConn, err = amqp.DialTLS(i.Config.LogsAmqpURI, cfg)
+		amqpConn, err = amqp.DialConfig(i.Config.LogsAmqpURI,
+			amqp.Config{
+				Heartbeat:       i.Config.AmqpHeartbeat,
+				Locale:          "en_US",
+				TLSClientConfig: cfg,
+			})
 	} else if i.Config.AmqpInsecure {
-		amqpConn, err = amqp.DialTLS(
+		amqpConn, err = amqp.DialConfig(
 			i.Config.LogsAmqpURI,
-			&tls.Config{InsecureSkipVerify: true},
-		)
+			amqp.Config{
+				Heartbeat:       i.Config.AmqpHeartbeat,
+				Locale:          "en_US",
+				TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+			})
 	} else {
-		amqpConn, err = amqp.Dial(i.Config.LogsAmqpURI)
+		amqpConn, err = amqp.DialConfig(i.Config.LogsAmqpURI,
+			amqp.Config{
+				Heartbeat: i.Config.AmqpHeartbeat,
+				Locale:    "en_US",
+			})
 	}
 	if err != nil {
 		i.logger.WithField("err", err).Error("couldn't connect to the logs AMQP server")
-		return err
+		return nil, err
 	}
 
 	go i.amqpErrorWatcher(amqpConn)
 	i.logger.Debug("connected to the logs AMQP server")
-	return nil
+
+	logWriterFactory, err := NewAMQPLogWriterFactory(amqpConn, i.Config.RabbitMQSharding)
+	if err != nil {
+		return nil, err
+	}
+	return logWriterFactory, nil
 }
 
 func (i *CLI) amqpErrorWatcher(amqpConn *amqp.Connection) {