+++ image/env_selector.go
@@ -1,60 +1,35 @@
 package image
 
 import (
-	"fmt"
-	"regexp"
 	"strings"
 
 	"github.com/travis-ci/worker/config"
-)
-
-var (
-	nonAlphaNumRegexp = regexp.MustCompile(`[^a-zA-Z0-9_]+`)
 )
 
 // EnvSelector implements Selector for environment-based mappings
 type EnvSelector struct {
 	c *config.ProviderConfig
 
-	imageAliases map[string]string
+	lookup map[string]string
 }
 
 // NewEnvSelector builds a new EnvSelector from the given *config.ProviderConfig
 func NewEnvSelector(c *config.ProviderConfig) (*EnvSelector, error) {
 	es := &EnvSelector{c: c}
-	err := es.buildImageAliasMap()
-	if err != nil {
-		return nil, err
-	}
+	es.buildLookup()
 	return es, nil
 }
 
-func (es *EnvSelector) buildImageAliasMap() error {
-	aliasNames := es.c.Get("IMAGE_ALIASES")
-
-	aliasNamesSlice := strings.Split(aliasNames, ",")
-
-	imageAliases := map[string]string{}
+func (es *EnvSelector) buildLookup() {
+	lookup := map[string]string{}
 
 	es.c.Each(func(key, value string) {
 		if strings.HasPrefix(key, "IMAGE_") {
-			imageAliases[strings.ToLower(strings.Replace(key, "IMAGE_", "", -1))] = value
+			lookup[strings.ToLower(strings.Replace(key, "IMAGE_", "", -1))] = value
 		}
 	})
 
-	for _, aliasName := range aliasNamesSlice {
-		normalizedAliasName := strings.ToUpper(string(nonAlphaNumRegexp.ReplaceAll([]byte(aliasName), []byte("_"))))
-
-		key := fmt.Sprintf("IMAGE_ALIAS_%s", normalizedAliasName)
-		if !es.c.IsSet(key) {
-			return fmt.Errorf("missing config key %q", key)
-		}
-
-		imageAliases[aliasName] = es.c.Get(key)
-	}
-
-	es.imageAliases = imageAliases
-	return nil
+	es.lookup = lookup
 }
 
 func (es *EnvSelector) Select(params *Params) (string, error) {
@@ -65,16 +40,16 @@ func (es *EnvSelector) Select(params *Params) (string, error) {
 			continue
 		}
 
-		if s, ok := es.imageAliases[key]; ok {
+		if s, ok := es.lookup[key]; ok {
 			imageName = s
 			break
 		}
 	}
 
-	if selected, ok := es.imageAliases[imageName]; ok {
+	// check for one level of indirection
+	if selected, ok := es.lookup[imageName]; ok {
 		return selected, nil
 	}
-
 	return imageName, nil
 }
 
@@ -82,67 +57,60 @@ func (es *EnvSelector) buildCandidateKeys(params *Params) []string {
 	fullKey := []string{}
 	candidateKeys := []string{}
 
-	addKey := func(key string, addToFull bool) {
-		if addToFull {
-			fullKey = append(fullKey, key)
-		}
-		candidateKeys = append(candidateKeys, key)
-	}
-
 	hasLang := params.Language != ""
-	needsOS := true
-	needsLangSuffix := false
+	hasDist := params.Dist != ""
+	hasGroup := params.Group != ""
+	hasOS := params.OS != ""
 
-	if params.OS == "osx" && params.OsxImage != "" && hasLang {
-		addKey("osx_image_"+params.OsxImage, true)
-		needsOS = false
-		needsLangSuffix = true
-	}
-
-	if params.Dist != "" && hasLang {
-		addKey("dist_"+params.Dist, true)
-		needsLangSuffix = true
-	}
-
-	if params.Group != "" && hasLang {
-		addKey("group_"+params.Group+"_"+params.Language, false)
-		addKey("group_"+params.Group, true)
-		needsLangSuffix = true
-	}
-
-	if params.OS != "" && hasLang {
-		if needsOS {
-			addKey(params.OS, true)
-			addKey("os_"+params.OS, false)
+	if params.OS == "osx" && params.OsxImage != "" {
+		if hasLang {
+			candidateKeys = append(candidateKeys, "osx_image_"+params.OsxImage+"_"+params.Language)
 		}
-		needsLangSuffix = true
+		candidateKeys = append(candidateKeys, "osx_image_"+params.OsxImage)
+	}
+
+	if hasDist && hasGroup && hasLang {
+		candidateKeys = append(candidateKeys, "dist_"+params.Dist+"_group_"+params.Group+"_"+params.Language)
+		candidateKeys = append(candidateKeys, params.Dist+"_"+params.Group+"_"+params.Language)
+	}
+
+	if hasDist && hasLang {
+		candidateKeys = append(candidateKeys, "dist_"+params.Dist+"_"+params.Language)
+		candidateKeys = append(candidateKeys, params.Dist+"_"+params.Language)
+	}
+
+	if hasGroup && hasLang {
+		candidateKeys = append(candidateKeys, "group_"+params.Group+"_"+params.Language)
+		candidateKeys = append(candidateKeys, params.Group+"_"+params.Language)
+	}
+
+	if hasOS && hasLang {
+		candidateKeys = append(candidateKeys, "os_"+params.OS+"_"+params.Language)
+		candidateKeys = append(candidateKeys, params.OS+"_"+params.Language)
+	}
+
+	if hasDist {
+		candidateKeys = append(candidateKeys, "default_dist_"+params.Dist)
+		candidateKeys = append(candidateKeys, "dist_"+params.Dist)
+		candidateKeys = append(candidateKeys, params.Dist)
+	}
+
+	if hasGroup {
+		candidateKeys = append(candidateKeys, "default_group_"+params.Group)
+		candidateKeys = append(candidateKeys, "group_"+params.Group)
+		candidateKeys = append(candidateKeys, params.Group)
 	}
 
 	if hasLang {
-		if needsLangSuffix {
-			addKey(params.Language, true)
-			addKey("language_"+params.Language, false)
-		} else {
-			addKey("language_"+params.Language, true)
-		}
+		candidateKeys = append(candidateKeys, "language_"+params.Language)
+		candidateKeys = append(candidateKeys, params.Language)
+		candidateKeys = append(candidateKeys, params.Language)
 	}
 
-	if params.OS == "osx" && params.OsxImage != "" {
-		addKey("osx_image_"+params.OsxImage, false)
-	}
-
-	if params.Dist != "" {
-		addKey("dist_"+params.Dist, false)
-	}
-
-	if params.Group != "" {
-		addKey("group_"+params.Group, false)
-	}
-
-	if params.OS != "" {
-		addKey(params.OS, false)
-		addKey("default_"+params.OS, false)
-		addKey(params.OS, false)
+	if hasOS {
+		candidateKeys = append(candidateKeys, "default_os_"+params.OS)
+		candidateKeys = append(candidateKeys, "os_"+params.OS)
+		candidateKeys = append(candidateKeys, params.OS)
 	}
 
 	return append([]string{strings.Join(fullKey, "_")}, candidateKeys...)