+++ step_write_worker_info.go
@@ -29,5 +29,4 @@ func (s *stepWriteWorkerInfo) Run(state multistep.StateBag) multistep.StepAction
 	return multistep.ActionContinue
 }
 
-func (s *stepWriteWorkerInfo) Cleanup(state multistep.StateBag) {
-}
+func (s *stepWriteWorkerInfo) Cleanup(state multistep.StateBag) {}